# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/houseboats/{houseboatId}/media": {
      "get": {
        "operationId": "MediaController_list",
        "parameters": [
          {
            "name": "houseboatId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cabinId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Media"
        ]
      }
    },
    "/api/houseboats/{houseboatId}/media/images": {
      "post": {
        "operationId": "MediaController_uploadImage",
        "parameters": [
          {
            "name": "houseboatId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadImageDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Media"
        ]
      }
    },
    "/api/houseboats/{houseboatId}/media/videos": {
      "post": {
        "operationId": "MediaController_createVideo",
        "parameters": [
          {
            "name": "houseboatId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateVideoDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Media"
        ]
      }
    },
    "/api/houseboats/{houseboatId}/media/{mediaId}": {
      "delete": {
        "operationId": "MediaController_remove",
        "parameters": [
          {
            "name": "houseboatId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "mediaId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Media"
        ]
      }
    },
...
    "/api/houseboats/{houseboatId}/maintenance": {
      "get": {
        "operationId": "MaintenanceController_summary",
        "parameters": [
          {
            "name": "houseboatId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Maintenance"
        ]
      }
    },
    "/api/houseboats/{houseboatId}/maintenance/tasks": {
      "post": {
        "operationId": "MaintenanceController_createTask",
        "parameters": [
          {
            "name": "houseboatId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMaintenanceTaskDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Maintenance"
        ]
      }
    },
    "/api/houseboats/{houseboatId}/maintenance/tasks/{taskId}": {
      "patch": {
        "operationId": "MaintenanceController_updateTask",
        "parameters": [
          {
            "name": "houseboatId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "taskId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMaintenanceTaskDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Maintenance"
        ]
      }
    },
    "/api/houseboats/{houseboatId}/maintenance/tasks/{taskId}/complete": {
      "post": {
        "operationId": "MaintenanceController_completeTask",
        "parameters": [
          {
            "name": "houseboatId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "taskId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteTaskDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Maintenance"
        ]
      }
    },
    "/api/houseboats/{houseboatId}/maintenance/service-logs": {
      "post": {
        "operationId": "MaintenanceController_addServiceLog",
        "parameters": [
          {
            "name": "houseboatId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateServiceLogDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Maintenance"
        ]
      }
    },
    "/api/houseboats/{houseboatId}/maintenance/damage": {
      "get": {
        "operationId": "MaintenanceController_listDamage",
        "parameters": [
          {
            "name": "houseboatId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Maintenance"
        ]
      },
      "post": {
        "operationId": "MaintenanceController_reportDamage",
        "parameters": [
          {
            "name": "houseboatId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDamageDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Maintenance"
        ]
      }
    },
    "/api/houseboats/{houseboatId}/maintenance/damage/{damageId}": {
      "patch": {
        "operationId": "MaintenanceController_updateDamage",
        "parameters": [
          {
            "name": "houseboatId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "damageId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDamageDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Maintenance"
        ]
      }
    },
    "/api/houseboats/{houseboatId}/maintenance/engine-hours": {
      "post": {
        "operationId": "MaintenanceController_setEngineHours",
        "parameters": [
          {
            "name": "houseboatId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetEngineHoursDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Maintenance"
        ]
      }
    }
  },
  "info": {
    "title": "Houseboat API",
    "description": "Booking SaaS backend",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "hb_access"
      }
    },
    "schemas": {
...
      "UploadImageDto": {
        "type": "object",
        "properties": {
          "cabinId": {
            "type": "string"
          },
          "sortOrder": {
            "type": "number",
            "minimum": 0
          }
        }
      },
      "CreateVideoDto": {
        "type": "object",
        "properties": {
          "cabinId": {
            "type": "string"
          },
          "youtubeUrl": {
            "type": "string",
            "format": "uri"
          },
          "sortOrder": {
            "type": "number",
            "minimum": 0
          }
        },
        "required": [
          "youtubeUrl"
        ]
      },
...
      "CreateMaintenanceTaskDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 120
          },
          "intervalKind": {
            "type": "string",
            "enum": [
              "engine_hours",
              "calendar",
              "per_trip"
            ]
          },
          "intervalValue": {
            "type": "number",
            "description": "hours for engine_hours, days for calendar; ignored for per_trip.",
            "minimum": 1,
            "maximum": 1000000
          },
          "dueAtHours": {
            "type": "number",
            "minimum": 0,
            "maximum": 1000000
          },
          "dueDate": {
            "type": "string"
          },
          "notes": {
            "type": "string",
            "maxLength": 500
          }
        },
        "required": [
          "title",
          "intervalKind"
        ]
      },
      "UpdateMaintenanceTaskDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 120
          },
          "intervalValue": {
            "type": "number",
            "minimum": 1,
            "maximum": 1000000
          },
          "dueAtHours": {
            "type": "number",
            "minimum": 0,
            "maximum": 1000000
          },
          "dueDate": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "paused"
            ]
          },
          "notes": {
            "type": "string",
            "maxLength": 500
          }
        }
      },
      "CompleteTaskDto": {
        "type": "object",
        "properties": {
          "serviceDate": {
            "type": "string"
          },
          "engineHours": {
            "type": "number",
            "minimum": 0,
            "maximum": 1000000
          },
          "cost": {
            "type": "number",
            "minimum": 0
          },
          "note": {
            "type": "string",
            "maxLength": 500
          }
        }
      },
      "CreateServiceLogDto": {
        "type": "object",
        "properties": {
          "serviceDate": {
            "type": "string"
          },
          "engineHours": {
            "type": "number",
            "minimum": 0,
            "maximum": 1000000
          },
          "cost": {
            "type": "number",
            "minimum": 0
          },
          "note": {
            "type": "string",
            "maxLength": 500
          },
          "taskId": {
            "type": "string",
            "maxLength": 120
          }
        }
      },
      "CreateDamageDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 120
          },
          "detail": {
            "type": "string",
            "maxLength": 500
          }
        },
        "required": [
          "title"
        ]
      },
      "UpdateDamageDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "open",
              "fixed"
            ]
          },
          "repairCost": {
            "type": "number",
            "minimum": 0
          },
          "detail": {
            "type": "string",
            "maxLength": 500
          }
        }
      },
      "SetEngineHoursDto": {
        "type": "object",
        "properties": {
          "hours": {
            "type": "number",
            "minimum": 0,
            "maximum": 1000000
          }
        },
        "required": [
          "hours"
        ]
      }
    }
  }
}